requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appnope==0.1.4
asgiref==3.8.1
asteroid-filterbanks==0.4.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
DateTime==5.5
debugpy==1.8.14
decorator==5.2.1
decouple==0.0.7
distlib==0.3.9
Django==4.2.21
django-cors-headers==4.7.0
djangorestframework==3.16.0
docopt==0.6.2
drf-yasg==1.21.10
einops==0.8.1
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.12
ffmpeg-python==0.2.0
filelock==3.18.0
fonttools==4.56.0
frozenlist==1.6.0
fsspec==2025.3.2
future==1.0.0
h11==0.16.0
hdbscan==0.8.40
hf-xet==1.1.0
huggingface-hub==0.31.4
HyperPyYAML==1.2.2
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
inflection==0.5.1
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.0
julius==0.2.7
jupyter_client==8.6.3
jupyter_core==5.7.2
keybert==0.9.0
kiwisolver==1.4.7
lightning==2.5.1.post0
lightning-utilities==0.14.3
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.3
mysqlclient==2.2.0
nest-asyncio==1.6.0
networkx==2.6.1
numpy==1.22.4
omegaconf==2.3.0
optuna==4.3.0
packaging==25.0
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
pipenv==2025.0.1
platformdirs==4.3.7
primePy==1.3
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==6.30.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pyopenvidu==0.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-multipart==0.0.20
pytorch-lightning==2.5.1.post0
pytorch-metric-learning==2.8.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.11.0
semver==3.0.4
sentence-transformers==4.1.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
speechbrain==1.0.3
SQLAlchemy==2.0.40
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.2
sympy==1.14.0
tabulate==0.9.0
tensorboardX==2.6.2.2
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.7.0
torchmetrics==1.7.1
tornado==6.5
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
virtualenv==20.30.0
wcwidth==0.2.13
yarl==1.20.0
zipp==3.21.0
zope.interface==7.2
